Integrate contents of src/i18n.js to mailMerge.js
