+++ test/getKnownErrorMessages_.test.js
@@ -1,10 +1,9 @@
 const mailMerge = require('../src/mailMerge');
-const i18n = require('../src/i18n');
 
-Object.keys(i18n.MESSAGES).forEach((locale) => {
-  let matchArray = Object.keys(i18n.MESSAGES[locale])
+Object.keys(mailMerge.MESSAGES).forEach((locale) => {
+  let matchArray = Object.keys(mailMerge.MESSAGES[locale])
     .filter((messageKey) => messageKey.slice(0, 5) === 'error')
-    .map((key) => i18n.MESSAGES[locale][key]);
+    .map((key) => mailMerge.MESSAGES[locale][key]);
   test('getKnownErrorMessages_', () => {
     expect(mailMerge.getKnownErrorMessages_(locale)).toEqual(matchArray);
   });
